Encapsulate Attribute information in AttributeSet
