+++ backend_test/ProficiencyTestSet.cpp
@@ -10,21 +10,25 @@ protected:
 	virtual void setUp() {
 		set = ProficiencySet();
 	}
+
+	static const std::vector<Attribute> getAllAttributes() {
+		return DND::AttributeSet::getAllAttributes();
+	}
 };
 
 TEST_F(ProficiencySetTest, TestProficiencies) {
 
-	for each (Attribute attr in entity_details::ATTRIBUTE_TYPES) {
+	for each (Attribute attr in AttributeSet::getAllAttributes()) {
 		EXPECT_FALSE(set.hasProficiency(attr));
 	}
 
-	Attribute attr1 = entity_details::ATTRIBUTE_TYPES[0];
-	Attribute attr2 = entity_details::ATTRIBUTE_TYPES[1];
+	Attribute attr1 = getAllAttributes()[0];
+	Attribute attr2 = getAllAttributes()[1];
 
 	set.addProficiency(attr1);
 	set.addProficiency(attr2);
 
-	for each (Attribute attr in entity_details::ATTRIBUTE_TYPES) {
+	for each (Attribute attr in  getAllAttributes()) {
 		if (attr != attr1 && attr != attr2) 
 			EXPECT_FALSE(set.hasProficiency(attr));
 		else 
@@ -32,7 +36,7 @@ TEST_F(ProficiencySetTest, TestProficiencies) {
 	}
 
 	set.removeProficiency(attr1);
-	for each (Attribute attr in entity_details::ATTRIBUTE_TYPES) {
+	for each (Attribute attr in getAllAttributes()) {
 		if (attr != attr2)
 			EXPECT_FALSE(set.hasProficiency(attr));
 		else
@@ -45,8 +49,8 @@ TEST_F(ProficiencySetTest, TestBonusStats) {
 	EXPECT_EQ(set.getBonusStats(2), AttributeSet());
 	EXPECT_EQ(set.getBonusStats(5), AttributeSet());
 
-	Attribute attr1 = entity_details::ATTRIBUTE_TYPES[0];
-	Attribute attr2 = entity_details::ATTRIBUTE_TYPES[1];
+	Attribute attr1 = getAllAttributes()[0];
+	Attribute attr2 = getAllAttributes()[1];
 	set.addProficiency(attr1);
 	set.addProficiency(attr2);
 	AttributeSet bonus = set.getBonusStats(5);
@@ -54,7 +58,7 @@ TEST_F(ProficiencySetTest, TestBonusStats) {
 	EXPECT_EQ(bonus.getAttributeScore(attr1), 3);
 	EXPECT_EQ(bonus.getAttributeScore(attr2), 3);
 	
-	for each (Attribute attr in entity_details::ATTRIBUTE_TYPES) {
+	for each (Attribute attr in getAllAttributes()) {
 		if (attr != attr1 && attr != attr2)
 			EXPECT_EQ(bonus.getAttributeScore(attr), 0);
 	}
