--- dnd_backend/entities/Attribute.cpp
@@ -1,13 +1,0 @@
-#include "Attribute.h"
-
-namespace DND {
-	namespace entity_details {
-		const std::vector<Attribute> ATTRIBUTE_TYPES = { 
-			DND::Attribute::Strength,
-			DND::Attribute::Dexterity,
-			DND::Attribute::Constitution,
-			DND::Attribute::Intelligence,
-			DND::Attribute::Wisdom,
-			DND::Attribute::Charisma };
-	}
-}
