+++ dnd_backend/entities/AttributeSet.cpp
@@ -2,12 +2,14 @@
 #include <stdexcept>
 #include <string> 
 
+const std::vector<DND::Attribute> DND::AttributeSet::getAllAttributes() {
+	return ATTRIBUTES;
+}
 
 DND::AttributeSet::AttributeSet(): map(entity_details::ObjectCounter<Attribute>() ){}
 
-
 DND::AttributeSet::AttributeSet(int amt) : map(entity_details::ObjectCounter<Attribute>()) {
-	for each (Attribute attribute in entity_details::ATTRIBUTE_TYPES) {
+	for each (Attribute attribute in ATTRIBUTES) {
 		setAttribute(attribute, amt);
 	}
 }
@@ -20,7 +22,7 @@ DND::AttributeSet& DND::AttributeSet::operator+=(DND::AttributeSet const& other)
 DND::AttributeSet DND::AttributeSet::operator+(DND::AttributeSet const & other) const {
 	AttributeSet newSet(*this);
 
-	for each (Attribute attribute in entity_details::ATTRIBUTE_TYPES) {
+	for each (Attribute attribute in ATTRIBUTES) {
 		newSet.setAttribute(attribute, getAttributeScore(attribute) + other.getAttributeScore(attribute));
 	}
 
@@ -44,10 +46,19 @@ int DND::AttributeSet::getAttributeScore(DND::Attribute attr) const {
 }
 
 void DND::AttributeSet::setAttribute(DND::Attribute attr, int amt) {
-	if (amt < DND::MIN_ATTRIBUTE_VALUE || amt > MAX_ATTRIBUTE_VALUE) {
-		throw std::invalid_argument("Character attributes must be between " + std::to_string(MIN_ATTRIBUTE_VALUE)
-			+ " and " + std::to_string(MAX_ATTRIBUTE_VALUE));
+	if (amt < MIN_ATTRIBUTE_VALUE || amt > MAX_ATTRIBUTE_VALUE) {
+		throw std::invalid_argument("Character attributes must be between " + ::std::to_string(MIN_ATTRIBUTE_VALUE)
+			+ " and " + ::std::to_string(MAX_ATTRIBUTE_VALUE));
 	}
 
 	map.setAmount(attr, amt);
 }
+
+const std::vector<DND::Attribute> DND::AttributeSet::ATTRIBUTES = { 
+		DND::Attribute::Strength,
+		DND::Attribute::Dexterity,
+		DND::Attribute::Constitution,
+		DND::Attribute::Intelligence,
+		DND::Attribute::Wisdom,
+		DND::Attribute::Charisma
+};
