+++ dnd_backend/entities/SkillSet.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Attribute.h"
+#include "AttributeSet.h"
 #include "ProficiencySet.h"
 #include "ObjectCounter.h"
 #include <unordered_map>
